requirements.txt
apispec==5.2.1
attrs==21.4.0
Babel==2.10.1
blinker==1.4
boto3==1.23.0
botocore==1.26.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
Faker==13.11.0
Flask==2.1.2
Flask-Babel==2.0.0
Flask-BabelEx==0.9.4
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-Login==0.6.1
Flask-Mail==0.9.1
flask-mongoengine==1.0.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.4
flask-smorest==0.37.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
marshmallow==3.15.0
mongoengine==0.24.1
packaging==21.3
passlib==1.7.4
Pillow==9.1.0
pluggy==1.0.0
py==1.11.0
PyJWT==2.3.0
pymongo==4.1.1
pyparsing==3.0.8
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
speaklater==1.3
SQLAlchemy==1.4.36
tomli==2.0.1
urllib3==1.26.9
webargs==8.1.0
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0
